requirements.txt
torch
torchvision
pandas
Pillow
tqdm
numpy